+++ SPG.Application/SPG.Application.csproj
@@ -4,6 +4,7 @@
     <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
+    <RootNamespace>SPG.Application</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
@@ -12,7 +13,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\SPG.Data\SPG.Data.csproj" />
-    <ProjectReference Include="..\SPG.Intf\SPG.Intf.csproj" />
+    <ProjectReference Include="..\SPG.Domain\SPG.Domain.csproj" />
   </ItemGroup>
 
 </Project>
