+++ SPG.Domain/SPG.Domain.csproj
@@ -4,6 +4,7 @@
     <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
+    <RootNamespace>SPG.Domain</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
